if category relate to setmeal or dish, throw exception when try to delete category

+++ sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -3,13 +3,18 @@ package com.sky.service.impl;
 import com.github.pagehelper.Constant;
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.sky.constant.MessageConstant;
 import com.sky.constant.StatusConstant;
 import com.sky.context.BaseContext;
 import com.sky.dto.CategoryDTO;
 import com.sky.dto.CategoryPageQueryDTO;
 import com.sky.entity.Category;
 import com.sky.entity.Employee;
+import com.sky.exception.DeletionNotAllowedException;
+import com.sky.handler.GlobalExceptionHandler;
 import com.sky.mapper.CategoryMapper;
+import com.sky.mapper.DishMapper;
+import com.sky.mapper.SetMealMapper;
 import com.sky.result.PageResult;
 import com.sky.service.CategoryService;
 import org.springframework.beans.BeanUtils;
@@ -24,6 +29,10 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Autowired
     private CategoryMapper categoryMapper;
+    @Autowired
+    private SetMealMapper setMealMapper;
+    @Autowired
+    private DishMapper dishMapper;
     @Override
     public void update(CategoryDTO categoryDTO) {
 
@@ -35,11 +44,10 @@ public class CategoryServiceImpl implements CategoryService {
 
     }
 
-    @Override
-    public List<Category> list(Integer type) {
-
-        return categoryMapper.list(type);
-    }
+//    @Override
+//    public List<Category> list(Integer type) {
+//        return categoryMapper.list(type);
+//    }
 
     @Override
     public PageResult page(CategoryPageQueryDTO categoryPageQueryDTO) {
@@ -64,6 +72,14 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void delete(Long id) {
+        Integer count = setMealMapper.countByCategoryId(id);
+        if(count > 0){
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_SETMEAL);
+        }
+        count = dishMapper.countDishByCategoryId(id);
+        if(count > 0){
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_DISH);
+        }
         categoryMapper.delete(id);
     }
 
@@ -75,5 +91,6 @@ public class CategoryServiceImpl implements CategoryService {
                 .build();
         categoryMapper.update(category);
     }
+
 
 }
